Ensure jobs are only run on minions
It'd probably be nice to store more metadata with the job itself, but
I'm really new to Gorp and I figure I'd discuss that with folks first.

This should work for the bare minimum case - we only want to run jobs on
machines with their role set to 'empire_minion'.

Also not sure how to make this work in the single vagrant box
environment.  I've added the metadata to make the Vagrant box an
empire_minion since we don't currently use empire_controller for
anything in the system, but I could see that changing.