Connects to etcd servers & looks for live nodes
This change allows us to specify that we only want a single etcd cluster
host, rather than all of them, in the output.  The big change that came
with that though is that we now verify that the etcd peers are alive and
get the list directly from the live etcd hosts (the first we find),
rather than just trusting the discoveryURL.

Oh, also, some pretty basic tests. First time really doing Go tests.