Use ZoneID when finding internal zone
This is so we don't accidentally update the wrong internal zone when
creating CNAMEs for load balancers when there are multiple zones that
match the internal zone name (empire.)

Also updates cloudformation/bootstrap to allow for non-us-east-1a AZs,
since some accounts won't have us-east-1a (mine, for example).

Adds output for HostedZoneID to the cloudformation.