Fix confusing error when no tasks present
Fixes #472

When a service doesn't yet exist for a new app, we would return a nasty
error message before. This removes that error message.

That said, this only returns a blank response now.  Not sure if there'd
be something better here.

When an app doesn't exist at all in Empire, we still get an appropriate
error:

```
error: Request failed, the specified resource does not exist
```