Add methods that will allow us to attach specific headers to a request
This will be used to add "Commit-Message" header for any action that
requires a commit message.

I went with <Method>WithHeaders to limit the surface area of the
changes. I didn't want to have to go back and adjust all the existing
calls to and pass a nil value for headers.