Simplifies scheduler.Run to only take a Manifest.
This commit simplifies the `twelvefactor.Scheduler.Run` interface method
to only take a `twelvefactor.Manifest` instead of both a
`twelvefactor.Manifest` and `twelvefactor.Process`. The manifest
contains the processes that should be run.

The commit also changes the `emp run` functionality, so that when a
"named" command is run, it will use the stored process configuration
(e.g. constraints) instead of the defaults.