Separate stdout/stderr streams for one-off commands.
This refactors the internal `empire.Empire.Run` method to support
separate I/O streams for stdout and stderr. Previously, stdout and
stderr from the attached container would be merged together.

This comment does not change the Heroku API to include support for
multiplexing the streams, so stdout/stderr will still be merged from the
clients perspective.