Allows behavior of image digest resolution to be controlled.
This adds a new --docker.digests flag (DOCKER_DIGESTS env var) that
allows the behavior of image resolution to be controlled. The default is
to use a digest if available, but can be changed to enforce digests
(error out if the image has no immutable digest), or disabled to get the
same behavior as before (no digest resolution).