Fix pull by digest in Docker scheduling backend
A previous change was made in fed3ad5a0b2d2b022b9599c69f115c9e54f994bc,
which allowed the registry operations to handle digests, however, that
change was no propagated through to the Docker scheduling backend. This
refactors the that change so that the same behavior can be applied to
both locations.

Fixes #1111