Fix emp run lockup in 0.13
In 0.13, stdout/stderr were split using the stdcopy package.
Unfortunately, this doesn't play well with streamhttp.Heartbeat, which
sends null characters and causes the emp client to lockup.

This removes the heartbeat, since it's mostly unecessary, and causes
other weird side effects (e.g. if you're using tmux, the pain will get
"active" because it thinks there's output).

If using an ELB in front of the Empire daemon, the best thing to do is
set a high idle timeout (maximum is 60 minutes) to prevent emp run
sessions with no output from being killed. This is already the default
recommendation in docs/cloudformation.json.