Ensures that the Custom::ECSService doesn't clobber DesiredCount on the service.
If you're using Custom::ECSService with application autoscaling, anytime
the Custom::ECSService resource is updated, the UpdateService call would
clobber the currently set DesiredCount on the service.

This changes the behavior such that DesiredCount is only updated on the
ECS service if DesiredCount is changed in the CloudFormation resource
properties. This makes it easier to use Custom::ECSService with auto
scaling.