Produce statically-linked emp and empire binaries
This change sets flags on the linker to ensure that all libraries are
statically linked. It also forces the compiler to use the Go
implementations of networking libraries (the default).

Making these changes allows the binaries to be run without libc in a
scratch Docker container, or in distributions like Alpine that don't
include libc.

Fixes #1068